requirements.txt
transformers
faiss-cpu
torch
torchvision
torchaudio
numpy 
pandas